+++ examples/single-region_complete-pam/main.tf
@@ -49,6 +49,15 @@ locals {
   pta_vm_name     = "PAMonCloud-TF-PTA"
   pta_vm_hostname = "PTA"
   pta_vm_size     = "Standard_D4s_v3"
+
+  # Bastion locals (used when var.deploy_bastion is true):
+  bastion_vm_name         = "PAMonCloud-TF-Bastion"
+  bastion_vm_hostname     = "bastion"
+  bastion_vm_size         = "Standard_D2as_v7"
+  bastion_image_publisher = "MicrosoftWindowsServer"
+  bastion_image_offer     = "WindowsServer"
+  bastion_image_sku       = "2022-datacenter-g2"
+  bastion_image_version   = "latest"
 }
 
 provider "azurerm" {
@@ -66,11 +75,36 @@ module "pam-network" {
   vnet_cidr_primary          = local.vnet_cidr_primary
   users_access_cidr          = local.users_access_cidr
   administrative_access_cidr = local.administrative_access_cidr
+  bastion_access_cidr        = var.bastion_access_cidr
 }
 
 ################################################################################
-# vault Module
+# bastion Module (optional)
 ################################################################################
+module "bastion" {
+  count  = var.deploy_bastion ? 1 : 0
+  source = "../../modules/bastion"
+
+  vm_name             = local.bastion_vm_name
+  vm_hostname         = local.bastion_vm_hostname
+  vm_size             = local.bastion_vm_size
+  resource_group_name = module.pam-network.rg_name
+  location            = module.pam-network.vnet_location.primary
+  availability_zone   = []
+  subnet_id           = module.pam-network.public_subnet_id.primary
+  vm_admin_user       = local.vm_admin_user
+  vm_admin_password   = var.vm_admin_password
+  image_publisher     = local.bastion_image_publisher
+  image_offer         = local.bastion_image_offer
+  image_sku           = local.bastion_image_sku
+  image_version       = local.bastion_image_version
+
+  depends_on = [module.pam-network]
+}
+
+# ##############################################################################
+# vault Module
+# ##############################################################################
 module "vault_vm" {
   source = "../../modules/vault"