+++ modules/component/versions.tf
@@ -1,5 +1,5 @@
 terraform {
-  required_version = "1.9.8"
+  required_version = ">= 1.9.8, <= 1.13.5"
   required_providers {
     azurerm = {
       source  = "hashicorp/azurerm"